%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  discretize.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{addCubeCol.Discretize}
\alias{addCubeCol.Discretize}
\alias{Discretize.addCubeCol}
\alias{addCubeCol,Discretize-method}

\title{Adds a 2D cube with color to the plot}

\description{
  Internal function.
}

\usage{
\method{addCubeCol}{Discretize}(this, cube, color=NULL, ...)
}

\arguments{
 \item{cube}{ The cube under consideration which is a (2x2) matrix containing the bounds of the variables (columnwise) (bivariate case).   }
 \item{color}{ Color of the cube. }
 \item{...}{Not used.}
}

\value{
  NULL
}

\author{Lars Relund \email{lars@relund.dk}}

\seealso{
  For more information see \code{\link{Discretize}}.
}


\keyword{internal}
\keyword{methods}
