%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  discretize.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{addPoints.Discretize}
\alias{addPoints.Discretize}
\alias{Discretize.addPoints}
\alias{addPoints,Discretize-method}

\title{Adds center points to the plot}

\description{
  Internal function.
}

\usage{
\method{addPoints}{Discretize}(this, cubes, ...)
}

\arguments{
 \item{cubes}{ The list of hypercubes. }
 \item{...}{Not used.}
}

\value{
  NULL
}

\author{Lars Relund \email{lars@relund.dk}}

\seealso{
  For more information see \code{\link{Discretize}}.
}


\keyword{internal}
\keyword{methods}
