%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  discretize.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{addText.Discretize}
\alias{addText.Discretize}
\alias{Discretize.addText}
\alias{addText,Discretize-method}

\title{Add text to the plot}

\description{
  Internal function.
}

\usage{
\method{addText}{Discretize}(this, cubes, text, ...)
}

\arguments{
 \item{cubes}{ The list of hypercubes. }
 \item{text}{ Text to be added to each hypercube.}
 \item{...}{Not used.}
}

\value{
  NULL
}

\author{Lars Relund \email{lars@relund.dk}}

\seealso{
  For more information see \code{\link{Discretize}}.
}


\keyword{internal}
\keyword{methods}
