%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  discretizeGaussian-package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{discretizeGaussian-package}
\alias{discretizeGaussian-package}
\alias{discretizeGaussian}
\docType{package}

\title{Package discretizeGaussian}


\description{
 Contain functions for discritizing an uni/bivariate normal distribution. Both uniform and nonuniform discretization possible. To get an overview over the methods implemented see \link{Discretize}.
}

\details{The package is implemented using R.oo and implements a single class \code{Discretize}. See the examples on how to use. }

\author{Lars Relund \email{lars@relund.dk}}

\references{

[1] Nielsen, L.R.; Jørgensen, E. & Højsgaard, S. Embedding a state space model into a Markov decision process Dept. of Genetics and Biotechnology, Aarhus University, 2008. \cr

[2] Nielsen, L.R.; Jørgensen, E.; Kristensen, A.R. & Østergaard, S. Optimal Replacement Policies for Dairy Cows Based on Daily Yield Measurements Dept. of Genetics and Biotechnology, Aarhus University, 2009. \cr
}
\keyword{package}
