%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  discretize.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{plotCubes.Discretize}
\alias{plotCubes.Discretize}
\alias{Discretize.plotCubes}
\alias{plotCubes,Discretize-method}

\title{Plot the cubes (only bivariate distributions)}

\description{
  Internal function.
}

\usage{
\method{plotCubes}{Discretize}(this, cubes, start, colors, ...)
}

\arguments{
 \item{cubes}{ The list of hypercubes. }
 \item{start}{ An cube used to set the plot area.}
 \item{colors}{ An integer vector of same length as the number of cubes used to give the cubes colors. The color is set by the integer value. }
 \item{...}{Further arguments passed to plot.}
}

\value{
  NULL
}

\author{Lars Relund \email{lars@relund.dk}}

\seealso{
  For more information see \code{\link{Discretize}}.
}


\keyword{internal}
\keyword{methods}
