%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  discretize.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{ratio.Discretize}
\alias{ratio.Discretize}
\alias{Discretize.ratio}
\alias{ratio,Discretize-method}

\title{Calc max divided by min density value}

\description{
  Internal function.
}

\usage{
\method{ratio}{Discretize}(this, x, mu, sigma, ...)
}

\arguments{
 \item{x}{ Values to calc. density for.  }
 \item{mu}{ The mean. }
 \item{sigma}{ The covariate matrix. }
 \item{len}{ The number of samples of each coordinate.}
 \item{...}{Not used.}
}

\value{
  Calc max divided by min density value.
}

\author{Lars Relund \email{lars@relund.dk}}

\seealso{
  For more information see \code{\link{Discretize}}.
}


\keyword{internal}
\keyword{methods}
