%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  discretize.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{volCube.Discretize}
\alias{volCube.Discretize}
\alias{Discretize.volCube}
\alias{volCube,Discretize-method}

\title{Volume/length of cube}

\description{
  Internal function.
}

\usage{
\method{volCube}{Discretize}(this, cube, ...)
}

\arguments{
 \item{cube}{ The cube under consideration which is a (2x2) matrix containing the bounds (columnwise) (bivariate case) or vector of length 2 (univariate case).   }
 \item{...}{Not used.}
}

\author{Lars Relund \email{lars@relund.dk}}

\references{
Based on
\emph{Kozlov, A. & Koller, D. Nonuniform dynamic discretization in hybrid networks The Thirteenth Conference on Uncertainty in Artificial Intelligence (UAI-97), 1997, 314-325  }}


\keyword{internal}
\keyword{methods}
